requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
elevenlabs==1.2.2
feedgen==1.0.0
feedparser==6.0.11
openai==1.107.0
httpx>=0.27.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.0
selenium==4.21.0
lxml==5.2.2
boto3==1.34.69
pytz==2024.1
streamlit==1.32.0
pydub==0.25.1
google-cloud-texttospeech==2.16.3